backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests>=2.32.3
beautifulsoup4>=4.12.3
google-api-python-client>=2.143.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
